+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and others.
 * Feature: Added debug print function special value, that includes source position and
 dynamic type name: `$DEBUG ...`.
 * Feature: Added `std:time:now`, `std:srand` and `std:rand`.
+* Feature: Introduced variable and field assignment
+operator syntax `.x += 10` and `o.x += 10`.
 
 0.6.1 (2020-06-04)
 ==================